requirements.txt
#python>=3.9
decorator==4.4.2
networkx==2.5.1
scapy==2.4.5
posix-ipc==1.0.5
pyelftools @ git+https://github.com/eliben/pyelftools.git
websockets==10.0
pydot==1.4.2
#libgraphviz-dev (from apt)
readline==6.2.4.1
pyroute2
#setcap cap_net_admin,cap_sys_admin,cap_dac_override+eip python3.9
python3-xlib==0.15
#libncurses-dev (from apt)
async-property==0.2.1